requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
google-cloud-secret-manager==2.18.1
google-cloud-sql-connector[asyncpg]==1.7.0
google-cloud-pubsub==2.19.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.40.0
vertexai==1.40.0
reportlab==4.0.8
pypdf==3.17.4
PyPDF2==3.0.1
python-dotenv==1.0.0
httpx==0.26.0
redis==5.0.1
pytest==7.4.4
pytest-asyncio==0.23.3